# Add optimized ReadAll for http responses