## What does this PR do?

1. When the body content length is known it's much faster to pre-allocated the entire buffer once.
2. When the length is unknown using `bytes.NewBuffer` + `io.Copy` is much faster.

Also, added benchmarks to prove the difference.

<!-- Type of change
Please label this PR with one of the following labels, depending on the scope of your change:
- Bug
- Enhancement
- Breaking change
- Deprecation
- Cleanup
- Docs
-->



<!-- Mandatory
Explain here the changes you made on the PR. Please explain the WHAT: patterns used, algorithms implemented, design architecture, message processing, etc.
-->

## Why is it important?

<!-- Mandatory
Explain here the WHY, or the rationale/motivation for the changes.
-->
When we start using this function in main products it will significantly improve the performance.

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to have the PR approved

List here all the items you have verified BEFORE sending this PR. Please DO NOT remove any item, striking through those that do not apply. (Just in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have added tests that prove my fix is effective or that my feature works

## Benchmark

```
goos: darwin
goarch: arm64
pkg: github.com/elastic/elastic-agent-libs/transport/httpcommon
BenchmarkReadAll/size:_100/unknown_length/io.ReadAll-10                       	    1000	      1232 ns/op	     576 B/op	       3 allocs/op
BenchmarkReadAll/size:_100/unknown_length/bytes.Buffer+io.Copy-10         	    1000	      1219 ns/op	     224 B/op	       4 allocs/op

BenchmarkReadAll/size:_100/known_length/io.ReadAll-10                     	    1000	      1275 ns/op	     576 B/op	       3 allocs/op
BenchmarkReadAll/size:_100/known_length/bytes.Buffer+io.Copy-10           	    1000	      1206 ns/op	     176 B/op	       3 allocs/op

BenchmarkReadAll/size:_102400/unknown_length/io.ReadAll-10                	    1000	     60858 ns/op	  514405 B/op	      24 allocs/op
BenchmarkReadAll/size:_102400/unknown_length/bytes.Buffer+io.Copy-10      	    1000	      9833 ns/op	  106640 B/op	       8 allocs/op

BenchmarkReadAll/size:_102400/known_length/io.ReadAll-10                  	    1000	     44423 ns/op	  514401 B/op	      24 allocs/op
BenchmarkReadAll/size:_102400/known_length/bytes.Buffer+io.Copy-10        	    1000	      9229 ns/op	  106592 B/op	       7 allocs/op

BenchmarkReadAll/size:_1048576/unknown_length/io.ReadAll-10               	    1000	    474003 ns/op	 5241202 B/op	      33 allocs/op
BenchmarkReadAll/size:_1048576/unknown_length/bytes.Buffer+io.Copy-10     	    1000	     81912 ns/op	 1048722 B/op	       8 allocs/op

BenchmarkReadAll/size:_1048576/known_length/io.ReadAll-10                 	    1000	    457693 ns/op	 5241196 B/op	      33 allocs/op
BenchmarkReadAll/size:_1048576/known_length/bytes.Buffer+io.Copy-10       	    1000	     80061 ns/op	 1048674 B/op	       7 allocs/op
PASS
ok  	github.com/elastic/elastic-agent-libs/transport/httpcommon	1.543s
```

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
- Relates https://github.com/elastic/beats/issues/36151