@cmacknz 2024-02-14
https://github.com/golang/go/issues/50774 appears to be the upstream Go issue tracking the inefficiency of `io.ReadAll` here. Linking it for future reference.

Quoting the explanation of what causes this:

> The underlying reason is that ReadAll() uses append to resize its buffer, which goes up by a factor of 1.25 (slightly more in latest Go, but asymptotically the same) each time.
> 
> One might measure these alternatives on various axes:
> 
> number of bytes allocated and freed
> max size of heap required to read a given size
> amount of copying while resizing
> percentage waste in returned buffer
>
> The current implementation is better on the last one, and much worse on all the others.
>
> append is designed to minimize space wasted at the end of the buffer